test: separate safety api contract integration coverage
