+++ README.md
@@ -32,6 +32,7 @@ The current execution posture is:
 - prompt API behavior now also has its own dedicated integration module instead of living only inside the broad health suite,
 - provider API behavior now also has its own dedicated integration module instead of living only inside the broad health suite,
 - retrieval API behavior now also has its own dedicated integration module instead of living only inside the broad health suite,
+- safety API behavior now also has its own dedicated integration module instead of living only inside the broad health suite,
 - the provider gateway currently routes only to documented stub providers,
 - provider policy exposes which runtime modes are supported and how unsupported modes are rejected,
 - provider activation readiness is now exposed through a dedicated rollout-readiness endpoint,
