+++ docs/architecture/system-overview.md
@@ -106,6 +106,9 @@ provider contract coverage can evolve independently of the broader platform heal
 Retrieval API behavior is also verified through its own dedicated integration module so
 retrieval contract coverage can evolve independently of the broader platform health suite.
 
+Safety API behavior is also verified through its own dedicated integration module so safety
+contract coverage can evolve independently of the broader platform health suite.
+
 ### Async Runtime
 
 - `src/app/services/async_runtime_status.py`
